Load OpenSeadragon in stimulus controller (#634)
