+++ sitemedia/webpack-stats-ci.json
@@ -4,7 +4,6 @@
   "assets": [],
   "chunks": {
     "main": [],
-    "openseadragon": [],
     "search": [],
     "stimulus_app": [],
     "submenu": [],
