+++ webpack.config.js
@@ -15,7 +15,6 @@ module.exports = (env, options) => ({
             "./sitemedia/esm/main.esm.js",
             "./sitemedia/scss/main.scss",
         ],
-        openseadragon: "./sitemedia/js/openseadragon.js",
         search: "./sitemedia/js/search.js",
         stimulus_app: "./sitemedia/js/stimulus_app.js",
         submenu: "./sitemedia/js/submenu.js",
